+++ src/test/scala/uk/gov/hmrc/govukfrontend/views/TemplateUnitSpec.scala
@@ -202,7 +202,7 @@ abstract class TemplateUnitSpec[T: Reads, C <: Template1[T, HtmlFormat.Appendabl
   }
 
   private lazy val fixturesDirs: Seq[File] = {
-    val dir         = s"/fixtures"
+    val dir         = s"/fixtures/govuk-frontend"
     val fixturesDir = Try(File(Resource.my.getUrl(dir)))
       .getOrElse(throw new RuntimeException(s"Test fixtures folder not found: $dir"))